Разделил код на два файла
