+++ src/index.js
@@ -4,4 +4,4 @@ import './js/pagination-base';
 import './js/loader';
 import './js/render-top-films';
 import './js/scroll';
-import './js/trailer'
+import './js/trailer';
